// frontend/src/app/models/customer.model.ts
export interface Customer {
  customerNumber?: number;
  firstName?: string;
  lastName?: string;
  dateOfBirth?: string;
  houseName?: string;
  houseNumber?: string;
  postcode?: string;
  phoneHome?: string;
  phoneMobile?: string;
  emailAddress?: string;
  createdAt?: string;
  updatedAt?: string;
}

// Made with Bob